io_utils: move NodeNameWithHandler to a more specific package
NodeNameWithHandler has nothing to do with io. Let us place it in a
slightly more relevant package.
