+++ tests/storage/storage.go
@@ -60,6 +60,7 @@ import (
 	"kubevirt.io/kubevirt/tests/libkubevirt"
 	"kubevirt.io/kubevirt/tests/libkubevirt/config"
 	"kubevirt.io/kubevirt/tests/libnet"
+	"kubevirt.io/kubevirt/tests/libnode"
 	"kubevirt.io/kubevirt/tests/libpod"
 	"kubevirt.io/kubevirt/tests/libstorage"
 	"kubevirt.io/kubevirt/tests/libvmifact"
@@ -163,7 +164,7 @@ var _ = SIGDescribe("Storage", func() {
 			}
 
 			BeforeEach(func() {
-				nodeName = tests.NodeNameWithHandler()
+				nodeName = libnode.NodeNameWithHandler()
 				address, device = CreateErrorDisk(nodeName)
 				pv, pvc, err = CreatePVandPVCwithFaultyDisk(nodeName, device, testsuite.GetTestNamespace(nil))
 				Expect(err).NotTo(HaveOccurred(), "Failed to create PV and PVC for faulty disk")
@@ -1334,7 +1335,7 @@ var _ = SIGDescribe("Storage", func() {
 			}
 
 			BeforeEach(func() {
-				nodeName = tests.NodeNameWithHandler()
+				nodeName = libnode.NodeNameWithHandler()
 				address, device = CreateSCSIDisk(nodeName, []string{})
 			})
