+++ tests/usb/usb.go
@@ -43,7 +43,7 @@ var _ = Describe("[Serial][sig-compute][USB] host USB Passthrough", Serial, deco
 		kv := libkubevirt.GetCurrentKv(virtClient)
 		config = kv.Spec.Configuration
 
-		nodeName := tests.NodeNameWithHandler()
+		nodeName := libnode.NodeNameWithHandler()
 		Expect(nodeName).ToNot(BeEmpty())
 
 		// Emulated USB devices only on c9s providers. Remove this when sig-compute 1.26 is the
